pyproject.toml
[project]
name = "dj-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.1",
    "django-allauth[headless,headless-spec]>=65.14.0",
    "django-cors-headers>=4.9.0",
    "django-ninja-extra>=0.30.8",
    "ninja-schema>=0.14.3",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.2.0",
    "django-extensions>=4.1",
    "djlint>=1.36.4",
    "factory-boy>=3.3.3",
]
